refactor($ui): Move the UI into a separate module
Closes #9
